+++ gimmik/kernels/ptx/config/sm80_double.json
