+++ gimmik/kernels/ptx/config/sm80_float.json
@@ -10,6 +10,62 @@
   ],
   "kernels": [
     {
+      "template": "cstream-ksplit",
+      "family": "sparse",
+      "block": [
+        64,
+        2,
+        1
+      ],
+      "width": 1,
+      "params": {
+        "csz": 32,
+        "preload_c": true
+      },
+      "descriptor": "cstream-ksplit/preload-c/k2-c32-x64"
+    },
+    {
+      "template": "cstream-ksplit-v2",
+      "family": "sparse",
+      "block": [
+        64,
+        2,
+        1
+      ],
+      "width": 2,
+      "params": {
+        "csz": 16,
+        "preload_c": true
+      },
+      "conditions": {
+        "all": [
+          {
+            "field": "dtype",
+            "in": [
+              "float"
+            ]
+          },
+          {
+            "field": "n",
+            "is_not": null
+          },
+          {
+            "field": "n",
+            "divisible_by": 2
+          },
+          {
+            "field": "aligne",
+            "is_not": null
+          },
+          {
+            "field": "aligne",
+            "divisible_by": 2
+          }
+        ]
+      },
+      "descriptor": "cstream-ksplit-v2/preload-c/k2-c16-x64"
+    },
+    {
       "template": "cstream",
       "family": "sparse",
       "block": [
@@ -18,132 +74,28 @@
         1
       ],
       "width": 1,
-      "descriptor": "cstream"
-    },
-    {
-      "template": "bstream",
-      "family": "sparse",
-      "block": [
-        128,
-        1,
-        1
-      ],
-      "width": 1,
-      "descriptor": "bstream"
-    },
-    {
-      "template": "bstream-msplit",
-      "family": "sparse",
-      "block": [
-        32,
-        4,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "bsz": 24
-      },
-      "descriptor": "bstream-msplit/m4-b24-x32"
-    },
-    {
-      "template": "bstream-msplit-v2",
-      "family": "sparse",
-      "block": [
-        32,
-        4,
-        1
-      ],
-      "width": 2,
-      "params": {
-        "bsz": 16
-      },
-      "conditions": {
-        "all": [
-          {
-            "field": "dtype",
-            "eq": "float"
-          },
-          {
-            "field": "n",
-            "is_not": null
-          },
-          {
-            "field": "n",
-            "divisible_by": 2
-          },
-          {
-            "field": "aligne",
-            "is_not": null
-          },
-          {
-            "field": "aligne",
-            "divisible_by": 2
-          }
-        ]
-      },
-      "descriptor": "bstream-msplit-v2/m4-b16-x32"
-    },
-    {
-      "template": "bstream-msplit",
-      "family": "sparse",
-      "block": [
-        64,
-        1,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "bsz": 32
-      },
-      "conditions": {
-        "all": [
-          {
-            "field": "beta_zero",
-            "eq": true
-          },
-          {
-            "field": "m",
-            "lte": 320
-          },
-          {
-            "field": "k_used",
-            "gte": 64
-          }
-        ]
-      },
-      "descriptor": "bstream-msplit/m1-b32-x64"
-    },
-    {
-      "template": "cstream-ksplit",
-      "family": "sparse",
-      "block": [
-        32,
-        2,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "csz": 24
-      },
-      "descriptor": "cstream-ksplit/k2-c24-x32"
+      "descriptor": "cstream/x128"
     },
     {
       "template": "cstream-ksplit-v2",
       "family": "sparse",
       "block": [
-        32,
+        64,
         2,
         1
       ],
       "width": 2,
       "params": {
-        "csz": 24
+        "csz": 24,
+        "preload_c": true
       },
       "conditions": {
         "all": [
           {
             "field": "dtype",
-            "eq": "float"
+            "in": [
+              "float"
+            ]
           },
           {
             "field": "n",
@@ -163,40 +115,95 @@
           }
         ]
       },
-      "descriptor": "cstream-ksplit-v2/k2-c24-x32"
+      "descriptor": "cstream-ksplit-v2/preload-c/k2-c24-x64"
+    },
+    {
+      "template": "bstream",
+      "family": "sparse",
+      "block": [
+        256,
+        1,
+        1
+      ],
+      "width": 1,
+      "descriptor": "bstream/x256"
+    },
+    {
+      "template": "bstream-msplit-v2",
+      "family": "sparse",
+      "block": [
+        64,
+        2,
+        1
+      ],
+      "width": 2,
+      "params": {
+        "bsz": 32,
+        "preload_c": true
+      },
+      "conditions": {
+        "all": [
+          {
+            "field": "dtype",
+            "in": [
+              "float"
+            ]
+          },
+          {
+            "field": "n",
+            "is_not": null
+          },
+          {
+            "field": "n",
+            "divisible_by": 2
+          },
+          {
+            "field": "aligne",
+            "is_not": null
+          },
+          {
+            "field": "aligne",
+            "divisible_by": 2
+          }
+        ]
+      },
+      "descriptor": "bstream-msplit-v2/preload-c/m2-b32-x64"
     },
     {
       "template": "cstream-ksplit",
       "family": "sparse",
       "block": [
-        32,
-        4,
+        64,
+        2,
         1
       ],
       "width": 1,
       "params": {
-        "csz": 20
+        "csz": 24,
+        "preload_c": true
       },
-      "conditions": {
-        "field": "k_used",
-        "gt": 500
-      },
-      "descriptor": "cstream-ksplit/k4-c20-x32"
+      "descriptor": "cstream-ksplit/preload-c/k2-c24-x64"
     },
     {
-      "template": "cstream-v2",
+      "template": "cstream-ksplit-v2",
       "family": "sparse",
       "block": [
-        128,
-        1,
+        64,
+        2,
         1
       ],
       "width": 2,
+      "params": {
+        "csz": 32,
+        "preload_c": true
+      },
       "conditions": {
         "all": [
           {
             "field": "dtype",
-            "eq": "float"
+            "in": [
+              "float"
+            ]
           },
           {
             "field": "n",
@@ -207,16 +214,16 @@
             "divisible_by": 2
           },
           {
-            "field": "k_used",
-            "lte": 100
+            "field": "aligne",
+            "is_not": null
           },
           {
             "field": "aligne",
-            "is_null_or_divisible_by": 2
+            "divisible_by": 2
           }
         ]
       },
-      "descriptor": "cstream-v2/x128"
+      "descriptor": "cstream-ksplit-v2/preload-c/k2-c32-x64"
     }
   ]
 }
