--- gimmik/kernels/ptx/config/sm90_fp32.json
@@ -1,233 +1,0 @@
-{
-  "schema": 1,
-  "cc": [
-    9,
-    0
-  ],
-  "ptx": [
-    8,
-    6
-  ],
-  "kernels": [
-    {
-      "template": "cstream-ksplit",
-      "block": [
-        32,
-        2,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "csz": 20
-      },
-      "descriptor": "cstream-ksplit/k2-c20-x32",
-      "family": "sparse"
-    },
-    {
-      "template": "bstream-msplit",
-      "block": [
-        32,
-        8,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "bsz": 24
-      },
-      "descriptor": "bstream-msplit/m8-b24-x32",
-      "family": "sparse"
-    },
-    {
-      "template": "cstream-ksplit",
-      "block": [
-        32,
-        4,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "csz": 24
-      },
-      "descriptor": "cstream-ksplit/k4-c24-x32",
-      "family": "sparse"
-    },
-    {
-      "template": "bstream-msplit",
-      "block": [
-        64,
-        2,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "bsz": 32
-      },
-      "descriptor": "bstream-msplit/m2-b32-x64",
-      "family": "sparse"
-    },
-    {
-      "template": "bstream",
-      "block": [
-        64,
-        1,
-        1
-      ],
-      "width": 1,
-      "descriptor": "bstream/x64",
-      "family": "sparse"
-    },
-    {
-      "template": "bstream-msplit-v2",
-      "block": [
-        32,
-        4,
-        1
-      ],
-      "width": 2,
-      "params": {
-        "bsz": 16
-      },
-      "conditions": {
-        "all": [
-          {
-            "field": "dtype",
-            "eq": "float"
-          },
-          {
-            "field": "n",
-            "is_not": null
-          },
-          {
-            "field": "n",
-            "divisible_by": 2
-          },
-          {
-            "field": "aligne",
-            "is_not": null
-          },
-          {
-            "field": "aligne",
-            "divisible_by": 2
-          }
-        ]
-      },
-      "descriptor": "bstream-msplit-v2/m4-b16-x32",
-      "family": "sparse"
-    },
-    {
-      "template": "cstream-ksplit-v2",
-      "block": [
-        32,
-        2,
-        1
-      ],
-      "width": 2,
-      "params": {
-        "csz": 24
-      },
-      "conditions": {
-        "all": [
-          {
-            "field": "dtype",
-            "eq": "float"
-          },
-          {
-            "field": "n",
-            "is_not": null
-          },
-          {
-            "field": "n",
-            "divisible_by": 2
-          },
-          {
-            "field": "aligne",
-            "is_not": null
-          },
-          {
-            "field": "aligne",
-            "divisible_by": 2
-          }
-        ]
-      },
-      "descriptor": "cstream-ksplit-v2/k2-c24-x32",
-      "family": "sparse"
-    },
-    {
-      "template": "bstream-msplit",
-      "block": [
-        32,
-        2,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "bsz": 32
-      },
-      "descriptor": "bstream-msplit/m2-b32-x32",
-      "family": "sparse"
-    },
-    {
-      "template": "bstream-msplit",
-      "block": [
-        64,
-        1,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "bsz": 24
-      },
-      "descriptor": "bstream-msplit/m1-b24-x64",
-      "family": "sparse"
-    },
-    {
-      "template": "bstream-msplit",
-      "block": [
-        32,
-        4,
-        1
-      ],
-      "width": 1,
-      "params": {
-        "bsz": 32
-      },
-      "descriptor": "bstream-msplit/m4-b32-x32",
-      "family": "sparse"
-    },
-    {
-      "template": "cstream-v2",
-      "block": [
-        128,
-        1,
-        1
-      ],
-      "width": 2,
-      "conditions": {
-        "all": [
-          {
-            "field": "dtype",
-            "eq": "float"
-          },
-          {
-            "field": "n",
-            "is_not": null
-          },
-          {
-            "field": "n",
-            "divisible_by": 2
-          },
-          {
-            "field": "k_used",
-            "lte": 100
-          },
-          {
-            "field": "aligne",
-            "is_null_or_divisible_by": 2
-          }
-        ]
-      },
-      "descriptor": "cstream-v2/x128",
-      "family": "sparse"
-    }
-  ]
-}
