+++ gimmik/kernels/ptx/dmma-asmem.mako
